[PR #46] SSL Bugs + others
Adds functionality to the client for dumping chain entries for specific domains
makes the in_memory option for the client actually work
Fixes the database effectively always being in-memory for the server by adding extra logic to write the database to disk when updates are made to the chain
Fixed ssl errors in https_server by grabbing the chain entry for us from the database instead of crafting a new certificate on the fly
For some reason, the deletion of the challenge lock was undone somehow. This re-does those undone changes to do them again.